Create logic for selecting place on board
Created a current turn variable to keep track of which player is
currently playing. Then on click, the turn is attempted. If the spot
has not been clicked (checked by a class of
'clicked'), then add text to that spot of the 'currentTurn', and add the
class 'clicked' to this element, and finally change to the next player.