Format index.html and add styles with scss
Added classes to game board along with some scss properties to allow
the board to always be seen as a square grid regardless of window size.
Re-arranged the reset button so that it is centered. Added min-height to
alerts class to stop game board from moving  as alerts were added.
Changed background color to dark and default text color to match primary
button color.