Add change password function
Added form to HTML to change password, hidden on page load. Once
signed in, a button added to the gameBoard container allows access
to the password change form view which also hides the game.
Added function for this in ui.js.
Added event listeners for the button to access the form as well as the
submit button. Added event handlers and ajax request. Added necessary UI
updates.