Add games won message to welcome message
By storing the games completed and looping through the array,
I checked against the win conditions for "X" since the player is
always X currently.

It works a little different, so I made a new function to specifically
check api games opposed to the one used for games in play.