Add ai.js file, refactor onAttemptTurn function
I added a ai.js file to handle the take turn when AI toggle = true.
I changed the 'this' in my code to 'event.target', and moved all needed
variables to take a turn to my 'store.js' file in order to allow
the user to switch to ai without keeping the ai function on the
same page.