Fix bug where ai was not storing move correctly
Randomized AI added .
This bug manifested itself as a tie function not working. After tracking
it back to the 'occupiedSpots' array, it was made clear that choices
were not being pushed to the correct position. This is because
it was still set to push the first element, even though I had added
randomized AI.