Refactor checkAiWins to also block wins
Add param to checkAiWins to decide which player it is checking for
win spots, so that if there is no win spot for the ai,
it will check for the opposite player using the same logic, and
if it finds a spot, it will block it. Created function to check
for the other player's token.